pyproject.toml
[project]
name = "niece-edu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "comtypes>=1.4.9",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pdfplumber>=0.11.5",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "pywin32>=308",
    "tabula-py>=2.10.0",
    "tencentcloud-sdk-python>=3.0.1312",
    "tqdm>=4.67.1",
]
